Change return type of Area to double

+++ src/Widget.App/Interfaces/IWidgetService.cs
@@ -1,9 +1,7 @@
-using Widget.App.Interfaces;
-
 namespace Widget.App.Interfaces
 {
   public interface IWidgetService
   {
-    int? Area(IWidget widget);
+    double? Area(IWidget widget);
   }
 }
